+++ tests/resources/malawi_extraction_polygons.gpkg
