+++ tests/resources/puglia_extraction_polygons.gpkg
