+++ tests/resources/test_optical_cube.nc
