+++ tests/resources/waterloo_extraction_polygons.gpkg
