+++ tests/resources/wc_extraction_dataset.gpkg
