+++ tests/tests_integration/test_backend.py
@@ -1,6 +1,5 @@
 from openeo_gfmap.backend import cdse_connection, vito_connection
 
-
 def test_vito_connection_auth():
     con = vito_connection()
     info = con.describe_account()
