+++ tests/tests_integration/test_cloud_masking.py
