+++ tests/tests_unit/test_patch_extractors.py
